+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [1.5.0]
 ### Updates
 - Created new `GenericCrossAccountTrustRule` in order to check for CrossAccount issues for not mapped resources.
+- Added documentation regarding the deprecation of `CrossAccountTrustRule`, `S3CrossAccountTrustRule`, `KMSKeyCrossAccountTrustRule`, `ElasticsearchDomainCrossAccountTrustRule` and `OpenSearchDomainCrossAccountTrustRule`.
 ### Fixes
 - Stopped using `_statement_as_list()` when retrieving statements in `CrossAccountCheckingRule` in favor of `statement_as_list()`.